requirements.txt
Django>=1.8.4
django-appconf==0.6
django-cors-headers==1.1.0
djangorestframework==3.2.3
drf-nested-routers==0.10.0
django-rest-swagger==0.3.4
django-address>=0.1.4
mongoengine>=0.8.3
